Add an option to re-run analysis for all 2-pass filters
As suggested here:
https://forum.shotcut.org/t/force-stabilization-analysis-for-all-clips/47938
