+++ src/docks/encodedock.cpp
@@ -22,12 +22,12 @@
 #include "dialogs/addencodepresetdialog.h"
 #include "dialogs/listselectiondialog.h"
 #include "dialogs/multifileexportdialog.h"
+#include "findanalysisfilterparser.h"
 #include "jobqueue.h"
 #include "jobs/encodejob.h"
 #include "mainwindow.h"
 #include "mltcontroller.h"
 #include "models/markersmodel.h"
-#include "qmltypes/qmlapplication.h"
 #include "qmltypes/qmlfilter.h"
 #include "settings.h"
 #include "shotcut_mlt_properties.h"
@@ -1477,68 +1477,6 @@ void EncodeDock::runMelt(const QString &target, int realtime)
         m_immediateJob->start();
     }
 }
-
-class FindAnalysisFilterParser : public Mlt::Parser
-{
-private:
-    QUuid m_uuid;
-    QList<Mlt::Filter> m_filters;
-
-public:
-    FindAnalysisFilterParser()
-        : Mlt::Parser()
-    {}
-
-    QList<Mlt::Filter> &filters() { return m_filters; }
-
-    int on_start_filter(Mlt::Filter *filter)
-    {
-        QString serviceName = filter->get("mlt_service");
-        if (serviceName == "loudness" || serviceName == "vidstab") {
-            // If the results property does not exist, empty, or file does not exist.
-            QString results = filter->get("results");
-            if (results.isEmpty()) {
-                if (serviceName == "vidstab") {
-                    // vidstab requires a filename, which is only available when using a project folder.
-                    QString filename = filter->get("filename");
-                    if (filename.isEmpty() || filename.endsWith("vidstab.trf")) {
-                        filename = QmlApplication::getNextProjectFile("stab");
-                    }
-                    if (!filename.isEmpty()) {
-                        filter->set("filename", filename.toUtf8().constData());
-                        m_filters << Mlt::Filter(*filter);
-
-                        // Touch file to prevent overwriting the same file
-                        QFile file(filename);
-                        file.open(QIODevice::WriteOnly);
-                        file.resize(0);
-                        file.close();
-                    }
-                } else {
-                    m_filters << Mlt::Filter(*filter);
-                }
-            }
-        }
-        return 0;
-    }
-    int on_start_producer(Mlt::Producer *) { return 0; }
-    int on_end_producer(Mlt::Producer *) { return 0; }
-    int on_start_playlist(Mlt::Playlist *) { return 0; }
-    int on_end_playlist(Mlt::Playlist *) { return 0; }
-    int on_start_tractor(Mlt::Tractor *) { return 0; }
-    int on_end_tractor(Mlt::Tractor *) { return 0; }
-    int on_start_multitrack(Mlt::Multitrack *) { return 0; }
-    int on_end_multitrack(Mlt::Multitrack *) { return 0; }
-    int on_start_track() { return 0; }
-    int on_end_track() { return 0; }
-    int on_end_filter(Mlt::Filter *) { return 0; }
-    int on_start_transition(Mlt::Transition *) { return 0; }
-    int on_end_transition(Mlt::Transition *) { return 0; }
-    int on_start_chain(Mlt::Chain *) { return 0; }
-    int on_end_chain(Mlt::Chain *) { return 0; }
-    int on_start_link(Mlt::Link *) { return 0; }
-    int on_end_link(Mlt::Link *) { return 0; }
-};
 
 void EncodeDock::enqueueAnalysis()
 {
